refactor: ipne-mvp0 を ipne にリネーム
- featuresディレクトリ: ipne-mvp0 → ipne
- ページファイル: IpneMvp0Page → IpnePage
- 画像アセット: ipne_mvp0_*.webp → ipne_*.webp
- ルートパス: /ipne-mvp0 → /ipne
- すべてのインポート、参照を更新
- テスト実行確認済み（7/7 passed）
