+++ src/App.tsx
@@ -17,8 +17,8 @@ const NonBrakeDescentPage = lazy(() => import('./pages/NonBrakeDescentPage'));
 const DeepSeaShooterPage = lazy(
   () => import(/* webpackChunkName: "DeepSeaShooterPage" */ './pages/DeepSeaShooterPage')
 );
-const IpneMvp0Page = lazy(
-  () => import(/* webpackChunkName: "IpneMvp0Page" */ './pages/IpneMvp0Page')
+const IpnePage = lazy(
+  () => import(/* webpackChunkName: "IpnePage" */ './pages/IpnePage')
 );
 
 // アプリケーションのルートコンテナ
@@ -147,7 +147,7 @@ const App: React.FC = () => {
                 <Route path="/maze-horror" element={<MazeHorrorPage />} />
                 <Route path="/non-brake-descent" element={<NonBrakeDescentPage />} />
                 <Route path="/deep-sea-shooter" element={<DeepSeaShooterPage />} />
-                <Route path="/ipne-mvp0" element={<IpneMvp0Page />} />
+                <Route path="/ipne" element={<IpnePage />} />
               </Routes>
             </main>
           </Suspense>
